@@ -178,8 +178,8 @@ class ZetaDropDownState<T> extends State<ZetaDropdown<T>> {
   final OverlayPortalController _overlayPortalController = OverlayPortalController();
 
   final _link = LayerLink();
-  final _menuKey = GlobalKey();
-  final _childKey = GlobalKey();
+  final GlobalKey<State<StatefulWidget>> _menuKey = GlobalKey();
+  final GlobalKey<State<StatefulWidget>> _childKey = GlobalKey();
   _MenuPosition _menuPosition = _MenuPosition.down;
 
   double? _menuSize;
